33558 Move Workbook conversion to WorkbookController Moved function and tests
